mobile: move deep link listener into separate hook
