+++ apps/tlon-mobile/src/hooks/useDeepLink.ts
@@ -1,7 +1,16 @@
+import { useNavigation } from '@react-navigation/native';
+import type { NavigationProp } from '@react-navigation/native';
+import { parseActiveTab } from '@tloncorp/shared';
 import { useCallback, useEffect, useState } from 'react';
+import { Alert } from 'react-native';
 import branch from 'react-native-branch';
 
+import { useShip } from '../contexts/ship';
+import { useWebViewContext } from '../contexts/webview/webview';
+import { inviteShipWithLure } from '../lib/hostingApi';
 import storage from '../lib/storage';
+import type { TabParamList } from '../types';
+import { trackError } from '../utils/posthog';
 
 type Lure = {
   lure: string | undefined;
@@ -108,3 +117,58 @@ export const useDeepLink = () => {
     clearLure,
   };
 };
+
+export function useDeepLinkListener(initialWer?: string) {
+  const navigation = useNavigation<NavigationProp<TabParamList>>();
+  const webviewContext = useWebViewContext();
+  const { ship } = useShip();
+  const { wer, lure, clearDeepLink } = useDeepLink();
+  const [gotoPath, setGotoPath] = useState(initialWer ?? '');
+
+  useEffect(() => {
+    if (wer) {
+      setGotoPath(wer);
+      clearDeepLink();
+    }
+  }, [wer, clearDeepLink]);
+
+  // If lure is present, invite it and mark as handled
+  useEffect(() => {
+    if (ship && lure) {
+      (async () => {
+        try {
+          await inviteShipWithLure({ ship, lure });
+          Alert.alert(
+            '',
+            'Your invitation to the group is on its way. It will appear in the Groups list.',
+            [
+              {
+                text: 'OK',
+                onPress: () => null,
+              },
+            ],
+            { cancelable: true }
+          );
+        } catch (err) {
+          console.error('Error inviting ship with lure:', err);
+          if (err instanceof Error) {
+            trackError(err);
+          }
+        }
+
+        clearDeepLink();
+      })();
+    }
+  }, [ship, lure, clearDeepLink]);
+
+  // If deep link clicked, broadcast that navigation update to the
+  // webview and mark as handled
+  useEffect(() => {
+    if (gotoPath) {
+      webviewContext.setGotoPath(gotoPath);
+      const tab = parseActiveTab(gotoPath) ?? 'Groups';
+      navigation.navigate(tab, { screen: 'Webview' });
+      clearDeepLink();
+    }
+  }, [gotoPath, clearDeepLink, webviewContext, navigation]);
+}
